refactor(shared): separate bedrock provider metadata
