+++ src/shared/api/index.ts
@@ -22,188 +22,7 @@ import { CLAUDE_SONNET_4_1M_SUFFIX, CLAUDE_SONNET_4_1M_TIERS } from "./providers
 export * from "./baseTypes"
 export * from "./providers/anthropic"
 export * from "./providers/claudeCode"
-
-// AWS Bedrock
-// https://docs.aws.amazon.com/bedrock/latest/userguide/conversation-inference.html
-export type BedrockModelId = keyof typeof bedrockModels
-export const bedrockDefaultModelId: BedrockModelId = "anthropic.claude-sonnet-4-20250514-v1:0"
-export const bedrockModels = {
-	"anthropic.claude-sonnet-4-20250514-v1:0:1m": {
-		maxTokens: 8192,
-		contextWindow: 1_000_000,
-		supportsImages: true,
-		supportsPromptCache: true,
-		inputPrice: 3.0,
-		outputPrice: 15.0,
-		cacheWritesPrice: 3.75,
-		cacheReadsPrice: 0.3,
-		tiers: CLAUDE_SONNET_4_1M_TIERS,
-	},
-	"anthropic.claude-sonnet-4-20250514-v1:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: true,
-		inputPrice: 3.0,
-		outputPrice: 15.0,
-		cacheWritesPrice: 3.75,
-		cacheReadsPrice: 0.3,
-	},
-	"anthropic.claude-opus-4-20250514-v1:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: true,
-		inputPrice: 15.0,
-		outputPrice: 75.0,
-		cacheWritesPrice: 18.75,
-		cacheReadsPrice: 1.5,
-	},
-	"anthropic.claude-opus-4-1-20250805-v1:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: true,
-		inputPrice: 15.0,
-		outputPrice: 75.0,
-		cacheWritesPrice: 18.75,
-		cacheReadsPrice: 1.5,
-	},
-	"amazon.nova-premier-v1:0": {
-		maxTokens: 10_000,
-		contextWindow: 1_000_000,
-		supportsImages: true,
-
-		supportsPromptCache: false,
-		inputPrice: 2.5,
-		outputPrice: 12.5,
-	},
-	"amazon.nova-pro-v1:0": {
-		maxTokens: 5000,
-		contextWindow: 300_000,
-		supportsImages: true,
-
-		supportsPromptCache: true,
-		inputPrice: 0.8,
-		outputPrice: 3.2,
-		// cacheWritesPrice: 3.2, // not written
-		cacheReadsPrice: 0.2,
-	},
-	"amazon.nova-lite-v1:0": {
-		maxTokens: 5000,
-		contextWindow: 300_000,
-		supportsImages: true,
-
-		supportsPromptCache: true,
-		inputPrice: 0.06,
-		outputPrice: 0.24,
-		// cacheWritesPrice: 0.24, // not written
-		cacheReadsPrice: 0.015,
-	},
-	"amazon.nova-micro-v1:0": {
-		maxTokens: 5000,
-		contextWindow: 128_000,
-		supportsImages: false,
-
-		supportsPromptCache: true,
-		inputPrice: 0.035,
-		outputPrice: 0.14,
-		// cacheWritesPrice: 0.14, // not written
-		cacheReadsPrice: 0.00875,
-	},
-	"anthropic.claude-3-7-sonnet-20250219-v1:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-
-		supportsPromptCache: true,
-		inputPrice: 3.0,
-		outputPrice: 15.0,
-		cacheWritesPrice: 3.75,
-		cacheReadsPrice: 0.3,
-	},
-	"anthropic.claude-3-5-sonnet-20241022-v2:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-
-		supportsPromptCache: true,
-		inputPrice: 3.0,
-		outputPrice: 15.0,
-		cacheWritesPrice: 3.75,
-		cacheReadsPrice: 0.3,
-	},
-	"anthropic.claude-3-5-haiku-20241022-v1:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: true,
-		inputPrice: 0.8,
-		outputPrice: 4.0,
-		cacheWritesPrice: 1.0,
-		cacheReadsPrice: 0.08,
-	},
-	"anthropic.claude-3-5-sonnet-20240620-v1:0": {
-		maxTokens: 8192,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 3.0,
-		outputPrice: 15.0,
-	},
-	"anthropic.claude-3-opus-20240229-v1:0": {
-		maxTokens: 4096,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 15.0,
-		outputPrice: 75.0,
-	},
-	"anthropic.claude-3-sonnet-20240229-v1:0": {
-		maxTokens: 4096,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 3.0,
-		outputPrice: 15.0,
-	},
-	"anthropic.claude-3-haiku-20240307-v1:0": {
-		maxTokens: 4096,
-		contextWindow: 200_000,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 0.25,
-		outputPrice: 1.25,
-	},
-	"deepseek.r1-v1:0": {
-		maxTokens: 8_000,
-		contextWindow: 64_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 1.35,
-		outputPrice: 5.4,
-	},
-	"openai.gpt-oss-120b-1:0": {
-		maxTokens: 8192,
-		contextWindow: 128_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.15,
-		outputPrice: 0.6,
-		description:
-			"A state-of-the-art 120B open-weight Mixture-of-Experts language model optimized for strong reasoning, tool use, and efficient deployment on large GPUs",
-	},
-	"openai.gpt-oss-20b-1:0": {
-		maxTokens: 8192,
-		contextWindow: 128_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.07,
-		outputPrice: 0.3,
-		description:
-			"A compact 20B open-weight Mixture-of-Experts language model designed for strong reasoning and tool use, ideal for edge devices and local inference.",
-	},
-} as const satisfies Record<string, ModelInfo>
+export * from "./providers/bedrock"
 
 // OpenRouter
 // https://openrouter.ai/models?order=newest&supported_parameters=tools
